requirements.txt
Flask==2.2.2
duckduckgo_search
newspaper3k